features structure and providers refactored
