+++ examples/angular/src/app/app.module.ts
@@ -1,13 +1,12 @@
 import { NgModule } from '@angular/core';
 import { BrowserModule } from '@angular/platform-browser';
-
-import { AppProviderModule } from '../business-logic/domain/app.provider.module';
 import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
 import { LatestAskStoriesComponent } from './story/latest-ask-stories/latest-ask-stories.component';
 import { ShowStoryComponent } from './story/show-story/story.component';
-import {DefaultNavigationService, NavigationService} from "./services/navigation.service";
-import {Router} from "@angular/router";
+import { DefaultNavigationService, NavigationService } from "./services/navigation.service";
+import { Router} from "@angular/router";
+import { HACKER_NEWS_PROVIDERS } from "../features/hacker-news-story/hacker-news-story.provider.module";
 
 @NgModule({
   declarations: [
@@ -16,7 +15,6 @@ import {Router} from "@angular/router";
     ShowStoryComponent
   ],
   imports: [
-    AppProviderModule,
     AppRoutingModule,
     BrowserModule,
   ],
@@ -25,7 +23,9 @@ import {Router} from "@angular/router";
       provide: NavigationService,
       deps: [Router],
       useFactory: (router: Router) => new DefaultNavigationService(router),
-    }
+    },
+    // Features
+    ...HACKER_NEWS_PROVIDERS
   ],
   bootstrap: [AppComponent]
 })
