+++ examples/angular/src/features/hacker-news-story/data/data-sources/hacker-news-story.network.data-source.ts
