+++ examples/angular/src/features/hacker-news-story/data/entities/hacker-news-item.entity.ts
