+++ examples/angular/src/features/hacker-news-story/data/mappers/hacker-news-story.mapper.ts
