+++ examples/angular/src/features/hacker-news-story/data/queries/hacker-news.query.ts
