+++ examples/angular/src/features/hacker-news-story/data/service/hacker-news.service.ts
