+++ examples/angular/src/features/hacker-news-story/domain/interactors/get-hacker-news-latest-ask-stories.interactor.ts
