+++ examples/angular/src/features/hacker-news-story/domain/interactors/get-hacker-news-story.interactor.ts
