+++ examples/angular/src/features/hacker-news-story/domain/mappers/hacker-news-story.mapper.ts
