+++ examples/angular/src/features/hacker-news-story/domain/models/hacker-news-story.model.ts
